requirements.txt
fastapi
uvicorn
numpy
groq
python-dotenv

